requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0
